Reworked Conquerable clan halls Fixed Trigger skills Fixed mobs of same clan do not help Fixed Dam over time skills leave in 1 second Fixed chain heal do not heal if target is self and is in oly Fixed skill mastery ids: 330 and 331 Fixed some visual bug with pet Fixed cursed weapon players show correct name Fixed Q00512 missing instance xml Two more quests in java Q00407 and Q00634 Added some missing admin spawns for hellbound Cleaned up some effect handlers Added missing buylists Added new admin command to debug geodata //geogrid Fixed some invisible walking mobs Cleaned up some spam loggers Reworked nevit blessing now it should pause when enter in peace zone
compile and copy paste reunion-core.jar reunion-scripts.jar and reunion-engine.jar in correct folders

+++ trunk/L2J_ReunionProject_Core/java/l2r/gameserver/model/ChanceCondition.java
@@ -150,7 +150,7 @@ public final class ChanceCondition
 		return skillIds;
 	}
 	
-	public boolean trigger(int event, int damage, byte element, boolean playable, L2Skill skill)
+	public boolean trigger(int event, double damage, byte element, boolean playable, L2Skill skill)
 	{
 		if (_pvpOnly && !playable)
 		{
